pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smart-hospital-allocator"
version = "1.0.0"
description = "OpenEnv RL environment: Smart Hospital Resource Allocator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask==3.0.3",
    "gunicorn==22.0.0",
    "gymnasium==0.29.1",
    "numpy==1.26.4",
    "openai>=1.0.0",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-flask>=1.3.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths  = ["tests"]
addopts    = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["core", "app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-flask>=1.3.0",
]
